requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
PyMySQL==1.1.0
python-dotenv==1.0.0
wikipedia==1.4.0
gunicorn==21.2.0

